Revert "Fix some uplink implant bugs. (#43501)"
This reverts commit 5afe3d7ea45776b7dee227f2272e96e80c4a1aa5, reversing
changes made to 17d0f27b128901549e886cd1592f864c3dbe945b.
