+++ Content.Server/Store/Components/CurrencyComponent.cs
@@ -1,7 +1,8 @@
 using Content.Shared.FixedPoint;
+using Content.Shared.Store;
 using Robust.Shared.Serialization.TypeSerializers.Implementations.Custom.Prototype.Dictionary;
 
-namespace Content.Shared.Store.Components;
+namespace Content.Server.Store.Components;
 
 /// <summary>
 /// Identifies a component that can be inserted into a store
