+++ Content.Shared/Implants/SharedSubdermalImplantSystem.Relays.cs
@@ -1,9 +1,9 @@
 using Content.Shared.Chat;
 using Content.Shared.IdentityManagement.Components;
 using Content.Shared.Implants.Components;
+using Content.Shared.Interaction;
 using Content.Shared.Interaction.Events;
 using Content.Shared.Mobs;
-using Content.Shared.Store;
 
 namespace Content.Shared.Implants;
 
@@ -12,14 +12,11 @@ public abstract partial class SharedSubdermalImplantSystem
     public void InitializeRelay()
     {
         SubscribeLocalEvent<ImplantedComponent, MobStateChangedEvent>(RelayToImplantEvent);
+        SubscribeLocalEvent<ImplantedComponent, AfterInteractUsingEvent>(RelayToImplantEvent);
         SubscribeLocalEvent<ImplantedComponent, SuicideEvent>(RelayToImplantEvent);
         SubscribeLocalEvent<ImplantedComponent, TransformSpeakerNameEvent>(RelayToImplantEvent);
         SubscribeLocalEvent<ImplantedComponent, TransformSpeechEvent>(RelayToImplantEvent);
         SubscribeLocalEvent<ImplantedComponent, SeeIdentityAttemptEvent>(RelayToImplantEvent);
-
-        // Ref relays, for when you need to write to the event!
-        SubscribeLocalEvent<ImplantedComponent, CurrencyInsertAttemptEvent>(RefRelayToImplantEvent);
-        SubscribeLocalEvent<ImplantedComponent, GetStoreEvent>(RefRelayToImplantEvent);
     }
 
     /// <summary>
@@ -39,26 +36,6 @@ public abstract partial class SharedSubdermalImplantSystem
             RaiseLocalEvent(implant, relayEv);
         }
     }
-
-    /// <summary>
-    /// Relays events from the implanted to the implant.
-    /// </summary>
-    private void RefRelayToImplantEvent<T>(Entity<ImplantedComponent> entity, ref T args) where T : notnull
-    {
-        if (!_container.TryGetContainer(entity, ImplanterComponent.ImplantSlotId, out var implantContainer))
-            return;
-
-        var relayEv = new ImplantRelayEvent<T>(args, entity);
-        foreach (var implant in implantContainer.ContainedEntities)
-        {
-            if (args is HandledEntityEventArgs { Handled: true })
-                return;
-
-            RaiseLocalEvent(implant, relayEv);
-        }
-
-        args = relayEv.Event;
-    }
 }
 
 /// <summary>
@@ -66,7 +43,7 @@ public abstract partial class SharedSubdermalImplantSystem
 /// </summary>
 public sealed class ImplantRelayEvent<T> where T : notnull
 {
-    public T Event;
+    public readonly T Event;
 
     public readonly EntityUid ImplantedEntity;
