+++ Resources/Prototypes/Entities/Objects/Misc/subdermal_implants.yml
@@ -155,7 +155,7 @@
     - Syndicate
 
 - type: entity
-  parent: BaseSubdermalImplant
+  parent: [ BaseSubdermalImplant, StorePresetUplink ]
   id: UplinkImplant
   name: uplink implant
   description: This implant lets the user access a hidden Syndicate uplink at will.
